Minor bugfixes - Added game shortcut to screenshots directory - Added preset shortcut to screenshots directory

+++ ENBManager.Modules/Shared/Services/ScreenshotManager.cs
@@ -56,7 +56,8 @@ namespace ENBManager.Modules.Shared.Services
         {
             _logger.Debug("Deleting screenshot directory");
 
-            Directory.Delete(directory, true);
+            if (Directory.Exists(directory))
+                Directory.Delete(directory, true);
         }
 
         #endregion
